delete snapshot tests
